Se cambia el nombre de la clase empleado por usuario
